+++ src/main/java/com/example/Ebanking/service/TransactionServiceIF.java
@@ -15,6 +15,6 @@ import java.util.List;
  */
 public interface TransactionServiceIF {
     public void saveTransaction(TransactionEntity transactionEntity);
-    public List<TransactionEntity>getTransactionByDate(Date start,Date end,int id);
+    public List<TransactionEntity>getTransactionByDate(Date start,Date end,double id);
     public TransactionEntity getTransactionByID(int id);
 }
